+++ DespachoDjango/apps/inventario/urls.py
@@ -34,5 +34,4 @@ urlpatterns = [
     #path('api/', include(router.urls)), #<-- Comente esta seccion porque la API no la estamos usando y no es necesaria para el proyecto -->
     path('exportar/', views.exportar_inventario, name='exportar_inventario'),
     path('configurar-reporte/', views.configurar_reporte, name='configurar_reporte'),
-    path('reporte_inventario/', views.reporte_inventario, name='reporte_inventario'), #<-- Comente esta seccion porque no era una tarea mia -->
 ]
