+++ DespachoDjango/apps/inventario/views.py
@@ -382,217 +382,3 @@ def exportar_inventario(request):
 
     wb.save(response)
     return response
-
-def reporte_inventario(request):
-    # Obtener filtros del request
-    nombre = request.GET.get('nombre', '')
-    categoria_id = request.GET.get('categoria')
-    stock_status = request.GET.get('stock_status')
-    fecha_filtro = request.GET.get('fecha_filtro')
-
-    # Filtrar productos activos
-    productos = Product.objects.filter(activo=True)
-
-    # Filtro de nombre
-    if nombre:
-        productos = productos.filter(name__icontains=nombre)
-
-    # Filtro de categoría
-    if categoria_id:
-        productos = productos.filter(categoria__id=categoria_id)
-
-    # Filtro de estado de stock
-    if stock_status:
-        productos = [p for p in productos if p.get_stock_status()[1] == stock_status]
-
-    # Filtro de fecha
-    if fecha_filtro == 'hoy':
-        productos = productos.filter(created_at__date=timezone.now().date())
-    elif fecha_filtro == 'ultimo_mes':
-        productos = productos.filter(created_at__gte=timezone.now() - timedelta(days=30))
-    elif fecha_filtro == 'ultimo_ano':
-        productos = productos.filter(created_at__gte=timezone.now() - timedelta(days=365))
-
-    # Crear lista de productos con stock y valores calculados
-    productos_inventario = []
-    total_valor_inventario = 0
-
-    for producto in productos:
-        stock_actual = producto.get_stock_actual()
-        valor_inventario = producto.price * stock_actual
-        total_valor_inventario += valor_inventario
-
-        productos_inventario.append({
-            'producto': producto,
-            'stock_actual': stock_actual,
-            'valor_inventario': valor_inventario,
-            'stock_status': producto.get_stock_status()
-        })
-
-    # Pasar datos y categorías para los filtros al template
-    categorias = Categoria.objects.all()
-
-    # Verificar si se solicita descargar el PDF
-    if 'descargar_pdf' in request.GET:
-        return generar_pdf_reporte_inventario(request, productos_inventario, total_valor_inventario, categorias)
-    
-    # Verificar si se solicita descargar el Excel
-    if 'descargar_excel' in request.GET:
-        return generar_excel_reporte_inventario(productos_inventario, total_valor_inventario)
-
-    # Renderizar el reporte en la vista HTML normal
-    return render(request, 'inventario/reporte_inventario.html', {
-        'productos_inventario': productos_inventario,
-        'total_valor_inventario': total_valor_inventario,
-        'categorias': categorias,
-    })
-
-
-def generar_pdf_reporte_inventario(request, productos_inventario, total_valor_inventario, categorias):
-    """Genera el PDF del reporte inventario a partir del template HTML"""
-
-    # Renderizar el template a HTML utilizando render_to_string
-    html = render_to_string('inventario/reporte_inventario_pdf.html', {
-        'productos_inventario': productos_inventario,
-        'total_valor_inventario': total_valor_inventario,
-        'categorias': categorias,
-    })
-
-    # Crear un buffer para almacenar el PDF generado
-    buffer = BytesIO()
-
-    # Usar xhtml2pdf para crear el PDF
-    pisa_status = pisa.CreatePDF(html, dest=buffer)
-
-    # Si hubo un error en la creación del PDF, mostrar un mensaje
-    if pisa_status.err:
-        return HttpResponse('Error generando el PDF', status=500)
-
-    # Configurar la respuesta para que sea un archivo PDF
-    buffer.seek(0)
-
-    # Forzar la descarga del PDF
-    response = HttpResponse(buffer, content_type='application/pdf')
-    response['Content-Disposition'] = 'attachment; filename="reporte_inventario.pdf"'
-    
-    return response
-
-def generar_excel_reporte_inventario(productos_inventario, total_valor_inventario):
-    # Crear un nuevo Workbook
-    wb = Workbook()
-    ws = wb.active
-    ws.title = "Reporte de Inventario"
-
-    # Agregar encabezados
-    headers = [
-        "Nombre del Producto", 
-        "Categoría", 
-        "Descripción", 
-        "Precio", 
-        "Stock Mínimo", 
-        "Stock Actual", 
-        "Estado del Stock", 
-        "Valor Total en Inventario"
-    ]
-    for col_num, header in enumerate(headers, start=1):
-        cell = ws.cell(row=1, column=col_num, value=header)
-        cell.font = Font(bold=True)
-
-    # Agregar datos de productos
-    for row_num, producto in enumerate(productos_inventario, start=2):
-        ws.cell(row=row_num, column=1, value=producto['producto'].name)
-        ws.cell(row=row_num, column=2, value=producto['producto'].categoria.nombre)
-        ws.cell(row=row_num, column=3, value=producto['producto'].description)
-        ws.cell(row=row_num, column=4, value=float(producto['producto'].price))
-        ws.cell(row=row_num, column=5, value=producto['producto'].stock_minimo)
-        ws.cell(row=row_num, column=6, value=producto['stock_actual'])
-        ws.cell(row=row_num, column=7, value=producto['stock_status'][0])  # Estado
-        ws.cell(row=row_num, column=8, value=float(producto['valor_inventario']))
-
-    # Agregar total del inventario al final
-    total_row = len(productos_inventario) + 2
-    ws.cell(row=total_row, column=1, value="Valor Total del Inventario")
-    ws.cell(row=total_row, column=8, value=float(total_valor_inventario))
-    ws.cell(row=total_row, column=1).font = Font(bold=True)
-
-    # Preparar respuesta HTTP con el archivo Excel
-    response = HttpResponse(content_type="application/vnd.openxmlformats-officedocument.spreadsheetml.sheet")
-    response['Content-Disposition'] = 'attachment; filename="reporte_inventario.xlsx"'
-
-    # Guardar el Workbook en el response
-    wb.save(response)
-    return response
-
-def reporte_general_inventario(): #Plantilla para realizar reporte general
-    productos = Product.objects.filter(activo=True)
-    reporte_data = []
-    total_valor_inventario = Decimal('0.00')
-
-    for producto in productos:
-        stock_actual = producto.get_stock_actual()
-        stock_status, stock_status_class = producto.get_stock_status()
-        valor_total_producto = producto.price * stock_actual
-
-        reporte_data.append({
-            'nombre': producto.name,
-            'categoria': producto.categoria.name,
-            'precio': producto.price,
-            'stock_minimo': producto.stock_minimo,
-            'stock_actual': stock_actual,
-            'estado_stock': stock_status,
-            'valor_total': valor_total_producto,
-        })
-
-        total_valor_inventario += valor_total_producto
-
-    return reporte_data, total_valor_inventario
-
-def reporte_por_categoria_inventario(categoria_id): #Plantilla para realizar reporte por categoria
-    categoria = Categoria.objects.get(id=categoria_id)
-    productos = categoria.productos.filter(activo=True)
-    reporte_data = []
-    total_valor_inventario = Decimal('0.00')
-
-    for producto in productos:
-        stock_actual = producto.get_stock_actual()
-        stock_status, stock_status_class = producto.get_stock_status()
-        valor_total_producto = producto.price * stock_actual
-
-        reporte_data.append({
-            'nombre': producto.name,
-            'descripcion': producto.description,
-            'precio': producto.price,
-            'stock_minimo': producto.stock_minimo,
-            'stock_actual': stock_actual,
-            'estado_stock': stock_status,
-            'valor_total': valor_total_producto,
-        })
-
-        total_valor_inventario += valor_total_producto
-
-    return reporte_data, total_valor_inventario
-
-def reporte_bajos_en_stock_inventario(): #Plantilla para realizar reporte bajos en stock
-    productos = Product.objects.filter(activo=True)
-    reporte_data = []
-    total_valor_inventario = Decimal('0.00')
-
-    for producto in productos:
-        stock_actual = producto.get_stock_actual()
-        if stock_actual < producto.stock_minimo * 0.30:  # Si el stock es menor al 30% del stock mínimo
-            stock_status, stock_status_class = producto.get_stock_status()
-            valor_total_producto = producto.price * stock_actual
-
-            reporte_data.append({
-                'nombre': producto.name,
-                'categoria': producto.categoria.name,
-                'precio': producto.price,
-                'stock_minimo': producto.stock_minimo,
-                'stock_actual': stock_actual,
-                'estado_stock': stock_status,
-                'valor_total': valor_total_producto,
-            })
-
-            total_valor_inventario += valor_total_producto
-
-    return reporte_data, total_valor_inventario
