+++ DespachoDjango/apps/reportes/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+class ReportesConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'apps.reportes'
+    verbose_name = 'reportes'
