+++ DespachoDjango/core/settings.py
@@ -37,6 +37,7 @@ LOCAL_APPS = [
     'apps.users.apps.UsersConfig',
     'apps.inventario.apps.InventarioConfig',
     'apps.ventas.apps.VentasConfig',
+    'apps.reportes.apps.ReportesConfig',
 ]
 
 INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + LOCAL_APPS + ["django_browser_reload"]
